PERCEPTION: address frame history PR comments

+++ gnc/sub8_perception/include/sub8_pcl/cv_tools.hpp
@@ -53,7 +53,7 @@ public:
 	int frames_available();
 
 	const std::string topic_name;
-	const unsigned int history_size;
+	const size_t history_size;
 
 private:
 	ros::NodeHandle nh;
